Move all sources under src/
