+++ packages/fuselage-hooks/src/__tests__/useAutoFocus.spec.ts
@@ -1,7 +1,10 @@
 import { useState, MutableRefObject } from 'react';
 
-import { runHooks } from '../.jest/helpers';
-import { useAutoFocus } from '../src';
+import { runHooks } from '../jestHelpers';
+
+import { useAutoFocus } from '..';
+
+type DummyElement = { focus: () => void };
 
 describe('useAutoFocus hook', () => {
   it('returns a ref', () => {
@@ -13,7 +16,7 @@ describe('useAutoFocus hook', () => {
   it('invokes focus', () => {
     const focus = jest.fn();
     runHooks(() => useAutoFocus(), [
-      (ref: MutableRefObject<object>) => {
+      (ref: MutableRefObject<DummyElement>) => {
         ref.current = { focus };
       },
     ]);
@@ -24,7 +27,7 @@ describe('useAutoFocus hook', () => {
   it('does not invoke focus if isFocused is false', () => {
     const focus = jest.fn();
     runHooks(() => useAutoFocus(false), [
-      (ref: MutableRefObject<object>) => {
+      (ref: MutableRefObject<DummyElement>) => {
         ref.current = { focus };
       },
     ]);
@@ -34,7 +37,7 @@ describe('useAutoFocus hook', () => {
 
   it('invokes focus if isFocused is toggled', () => {
     const focus = jest.fn();
-    runHooks<[MutableRefObject<object>, (x: boolean) => void]>(() => {
+    runHooks<[MutableRefObject<DummyElement>, (x: boolean) => void]>(() => {
       const [isFocused, setFocused] = useState(false);
       return [useAutoFocus(isFocused), setFocused];
     }, [
