+++ packages/fuselage-hooks/src/__tests__/useLazyRef.server.spec.ts
@@ -1,5 +1,6 @@
-import { runHooksOnServer } from '../.jest/helpers';
-import { useLazyRef } from '../src';
+import { runHooksOnServer } from '../jestHelpers';
+
+import { useLazyRef } from '..';
 
 describe('useLazyRef hook on server', () => {
   it('returns the computed value immediately', () => {
