+++ packages/fuselage-hooks/src/__tests__/useMutableCallback.server.spec.ts
@@ -2,8 +2,9 @@
  * @jest-environment node
  */
 
-import { runHooksOnServer } from '../.jest/helpers';
-import { useMutableCallback } from '../src';
+import { runHooksOnServer } from '../jestHelpers';
+
+import { useMutableCallback } from '..';
 
 describe('useMutableCallback hook on server', () => {
   it('returns a callback that invokes the mutable one', () => {
