+++ packages/fuselage-hooks/src/__tests__/useToggle.server.spec.ts
@@ -2,8 +2,9 @@
  * @jest-environment node
  */
 
-import { runHooksOnServer } from '../.jest/helpers';
-import { useToggle } from '../src';
+import { runHooksOnServer } from '../jestHelpers';
+
+import { useToggle } from '..';
 
 describe('useToggle hook on server', () => {
   it('has false value when an initial value is undefined', () => {
