+++ packages/fuselage-hooks/src/types/ResizeObserver.d.ts
